Add api of home screen offers,mostsearch and trim stars and also modify the salons screen to display these data

+++ lib/modules/home/models/trim_star_model.dart
@@ -3,7 +3,7 @@ class TrimStarModel {
   String name;
   String image;
   String cover;
-  int rate;
+  double rate;
   TrimStarModel({
     this.id,
     this.name,
@@ -18,7 +18,7 @@ class TrimStarModel {
       name: json['name'],
       image: json['image'],
       cover: json['cover'],
-      rate: json['rate'],
+      rate: (json['rate'] as int).toDouble(),
     );
   }
 }
